+++ src/main/java/models/works/DeleteService.java
@@ -10,8 +10,11 @@ public class DeleteService {
 
     public void delete(long workNo) {
 
-        boolean result = workDao.delete(workNo);
-        if(!result) {
+        if(!workDao.exists(workNo)) {
+            throw new WorkNotFoundException();
+        }
+
+        if(!workDao.delete(workNo)) {
             throw new WorkDeleteException();
         }
     }
