+++ src/main/java/models/works/SaveService.java
@@ -36,5 +36,7 @@ public class SaveService {
         if(workNo != null && !workNo.isBlank()) {
             work.setWorkNo(Long.parseLong(workNo));
         }
+
+        save(work);
     }
 }
