+++ src/main/resources/mybatis/mapper/WorkListMapper.xml
@@ -57,4 +57,7 @@
         </trim>
         ORDER BY regDt
     </select>
+    <select id="exists" parameterType="models.works.Work" resultType="int">
+        SELECT COUNT(*) FROM WorkList WHERE workNo=#{workNo}
+    </select>
 </mapper>
